LayoutTests/fast/page-color-sampling/color-sampling-fixed-descendant-in-relative-container.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=100 ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body, html {
            width: 100%;
            height: 100%;
            font-family: system-ui;
        }

        .relative {
            position: relative;
            z-index: 100;
            width: 100%;
            height: 100%;
        }

        header {
            position: fixed;
            top: 0;
            left: 0;
            width: 100%;
            height: 80px;
            background: rgb(255, 100, 0);
            color: white;
        }

        .tall {
            width: 10px;
            height: 2000px;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.ensurePresentationUpdate();
        edgeColors = await UIHelper.fixedContainerEdgeColors();
        shouldBeEqualToString("edgeColors.top", "rgb(255, 100, 0)");
        finishJSTest();
    });
    </script>
</head>
<body>
<div class="relative">
    <header></header>
</div>
<div class="tall"></div>
</body>
</html>
